% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/karexcel.r
\name{ASDATA.FRAME}
\alias{ASDATA.FRAME}
\title{Convert to Data Frame}
\usage{
ASDATA.FRAME(x)
}
\arguments{
\item{x}{An object to be converted into a data frame.}
}
\value{
A data frame.
}
\description{
This function converts the given object to a data frame.
}
\examples{
data <- ASDATA.FRAME(matrix(1:4, ncol = 2))
}
