% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/karexcel.r
\name{cluster.graph}
\alias{cluster.graph}
\title{Cluster a Graph and Extract Largest Component}
\usage{
cluster.graph(network)
}
\arguments{
\item{network}{A graph object.}
}
\value{
A list containing the largest connected component graph,
        node membership, and node importance data frame.
}
\description{
This function applies clustering to a graph and extracts the largest
connected component.
}
\examples{
\dontrun{
  # Assuming 'network' is a predefined graph object
  cluster.graph(network)
}
}
