% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/karexcel.r
\name{f6}
\alias{f6}
\title{Extract Token Information from Data Frame}
\usage{
f6(UPIP)
}
\arguments{
\item{UPIP}{A data frame containing columns 'token', 'lemma', and 'upos'
for tokens, their lemmatized forms, and POS tags respectively.}
}
\value{
A data frame with columns 'TOKEN', 'LEMMA', and 'TYPE', each
        representing token, its lemma, and POS tag.
}
\description{
This function extracts token, lemma, and part-of-speech (POS) tag information
from a given data frame and compiles them into a new data frame.
}
\examples{
data <- data.frame(token = c("running", "jumps"),
                   lemma = c("run", "jump"),
                   upos = c("VERB", "VERB"))
token_info <- f6(data)
}
