% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/karexcel.r
\name{fun.all.sums}
\alias{fun.all.sums}
\title{Apply Suffix Modifications to Persian Words}
\usage{
fun.all.sums(v, TYPE = TYPE.org)
}
\arguments{
\item{v}{A character vector of Persian words.}

\item{TYPE}{A vector of suffix types for modification.}
}
\value{
A modified character vector.
}
\description{
This function iteratively applies a series of suffix modifications to a vector of Persian words.
}
\examples{
\dontrun{
  words <- c("Persian text here")
  modified_words <- fun.all.sums(words, TYPE)
}
}
