% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/karexcel.r
\name{fun.one.sums}
\alias{fun.one.sums}
\title{General Persian Suffix Modification}
\usage{
fun.one.sums(v, type)
}
\arguments{
\item{v}{A character vector of Persian words.}

\item{type}{A character string representing the suffix type.}
}
\value{
A modified character vector.
}
\description{
This function modifies Persian words based on a specified suffix type.
}
\examples{
\dontrun{
  words <- c("Persian text here")
  modified_words <- fun.one.sums(words, "Persian text here")
}
}
