% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/karexcel.r
\name{FUNbigrams}
\alias{FUNbigrams}
\title{Extract Bigram Information and Count Frequency}
\usage{
FUNbigrams(tf.bigrams)
}
\arguments{
\item{tf.bigrams}{A data frame with bigram terms and their frequency.}
}
\value{
A data frame with columns for each word in the bigram and their frequency.
}
\description{
This function processes a data frame containing bigrams and their frequency,
and creates a new data frame with separated words and their frequencies.
}
\examples{
tf_bigrams <- data.frame(term = c("hello_world", "shiny_app"),
                         term_freq = c(3, 2))
bigram_info <- FUNbigrams(tf_bigrams)
}
