% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/karexcel.r
\name{fungi}
\alias{fungi}
\title{Persian Suffix Modification}
\usage{
fungi(v)
}
\arguments{
\item{v}{A character vector of Persian words.}
}
\value{
A modified character vector.
}
\description{
This function modifies Persian words ending with 'Persian text here' suffix.
}
\examples{
\dontrun{
  words <- c("Persian text here")
  modified_words <- fungi(words)
}
}
