% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/karexcel.r
\name{funmi}
\alias{funmi}
\title{Modify Persian Words Starting with 'Persian text here'}
\usage{
funmi(v)
}
\arguments{
\item{v}{A character vector of Persian words.}
}
\value{
A modified character vector.
}
\description{
This function modifies Persian words starting with the prefix 'Persian text here'.
}
\examples{
 \dontrun{
  words <- c("Persian text here")
  modified_words <- funmi(words)
}
}
