% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/karexcel.r
\name{LEMMA}
\alias{LEMMA}
\title{Persian Lemmatization}
\usage{
LEMMA(Y, TYPE = TYPE.org)
}
\arguments{
\item{Y}{A character vector of Persian words.}

\item{TYPE}{A vector of suffix types for modification.}
}
\value{
A vector of lemmatized Persian words.
}
\description{
This function performs lemmatization on a vector of Persian words.
}
\examples{
\dontrun{
  words <- c("Persian text here")
  lemmatized_words <- LEMMA(words, TYPE)
}
}
