% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/karexcel.r
\name{network.cor}
\alias{network.cor}
\title{Create and Plot a Correlation Network}
\usage{
network.cor(dt, Terms, threshold = 0.4, pl = TRUE)
}
\arguments{
\item{dt}{A document-term matrix.}

\item{Terms}{A vector of terms to check for correlation.}

\item{threshold}{A numeric threshold for correlation.}

\item{pl}{A logical value to plot the network or not.}
}
\value{
If `pl` is TRUE, a plot of the network; otherwise, a data frame of correlations.
}
\description{
This function creates a correlation network based on specified terms
and a threshold, and optionally plots it.
}
