% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/karexcel.r
\name{PMI}
\alias{PMI}
\title{Calculate Pointwise Mutual Information (PMI)}
\usage{
PMI(x)
}
\arguments{
\item{x}{A data frame with columns 'token' and 'doc_id'.}
}
\value{
A data frame of keywords with their PMI scores.
}
\description{
This function calculates the PMI for collocations in a given text data.
}
\examples{
data <- data.frame(token = c("word1", "word2"), doc_id = c(1, 1))
pmi_scores <- PMI(data)
}
