% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/karexcel.r
\name{ScaleWeight}
\alias{ScaleWeight}
\title{Scale a Numeric Vector}
\usage{
ScaleWeight(x, lambda)
}
\arguments{
\item{x}{A numeric vector.}

\item{lambda}{A numeric scaling factor.}
}
\value{
A scaled numeric vector.
}
\description{
This function scales a numeric vector by a specified lambda value.
}
\examples{
scaled_vector <- ScaleWeight(1:10, 2)
}
