% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/karexcel.r
\name{server}
\alias{server}
\title{Server Logic for MadanText Shiny Application}
\usage{
server(input, output)
}
\arguments{
\item{input}{List of Shiny inputs.}

\item{output}{List of Shiny outputs.}
}
\description{
This function contains the server-side logic for the MadanText application.
It handles user inputs, processes data, and creates outputs to be displayed
in the UI.
}
