% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/karexcel.r
\name{set.graph}
\alias{set.graph}
\title{Set Graph Attributes}
\usage{
set.graph(network)
}
\arguments{
\item{network}{A graph object.}
}
\value{
The graph object with updated attributes.
}
\description{
This function sets various attributes for a given graph object, including
vertex degree and edge width.
}
