% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/madantextnetwork.r
\docType{data}
\name{ui}
\alias{ui}
\title{User Interface for MadanText}
\format{
An object of class \code{shiny.tag.list} (inherits from \code{list}) of length 4.
}
\usage{
ui
}
\value{
A Shiny UI object.
}
\description{
This function creates a user interface for the MadanText Shiny application.
It includes various input and output widgets for file uploads, text input,
and visualization.
}
\keyword{datasets}
