% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/madantextnetwork.r
\name{ASDATA.FRAME}
\alias{ASDATA.FRAME}
\title{Convert to Data Frame}
\usage{
ASDATA.FRAME(x)
}
\arguments{
\item{x}{An object to be converted into a data frame.}
}
\value{
Returns a data frame with rows and columns corresponding to the
original object's structure. If `x` is a matrix, each column in the matrix
becomes a column in the data frame. If `x` is a list where all elements
are of the same length, each element of the list becomes a column in the
data frame. Attributes such as rownames, colnames, and dimnames (if any)
are preserved in the conversion.
}
\description{
This function converts the given object to a data frame.
}
\examples{
data <- ASDATA.FRAME(matrix(1:4, ncol = 2))
}
