% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/madantextnetwork.r
\name{Community.Detection.Plot}
\alias{Community.Detection.Plot}
\title{Plot Community Detection in a Graph}
\usage{
Community.Detection.Plot(network)
}
\arguments{
\item{network}{A graph object.}
}
\value{
A plot visualizing the graph with nodes colored according to their
community membership. The plot also displays the modularity score
as a sub-title, indicating the strength of the community structure.
}
\description{
This function applies community detection to a graph and plots the result.
}
\examples{
\dontrun{
  # Assuming 'network' is a predefined graph object
  # network <- make_graph("Zachary")
  Community.Detection.Plot(network)
}
}
