% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/madantextnetwork.r
\name{ScaleWeight}
\alias{ScaleWeight}
\title{Scale a Numeric Vector}
\usage{
ScaleWeight(x, lambda)
}
\arguments{
\item{x}{A numeric vector.}

\item{lambda}{A numeric scaling factor.}
}
\value{
A numeric vector where each element of the input vector 'x' is
divided by the scaling factor 'lambda'. This results in a scaled
version of the input vector.
}
\description{
This function scales a numeric vector by a specified lambda value.
}
\examples{
scaled_vector <- ScaleWeight(1:10, 2)
}
