% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/madantextnetwork.r
\name{server}
\alias{server}
\title{Server Logic for MadanText Shiny Application}
\usage{
server(input, output)
}
\arguments{
\item{input}{List of Shiny inputs.}

\item{output}{List of Shiny outputs.}
}
\value{
This function sets up the reactive environment and output elements in
the Shiny application. It does not return any value but modifies the Shiny app's
UI based on user inputs and reactive expressions. It returns a Shiny Server object.
}
\description{
This function contains the server-side logic for the MadanText application.
It handles user inputs, processes data, and creates outputs to be displayed
in the UI.
}
