% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/madantextnetwork.r
\name{set.graph}
\alias{set.graph}
\title{Set Graph Attributes}
\usage{
set.graph(network)
}
\arguments{
\item{network}{A graph object.}
}
\value{
The input graph object with added attributes: 'degree' for each vertex
and 'width' for each edge. These attributes enhance the graph's
visual representation and analytical capabilities.
}
\description{
This function sets various attributes for a given graph object, including
vertex degree and edge width.
}
